+++ CHANGELOG.md
@@ -1,4 +1,7 @@
 ```
+* Commit: 85de522 - v151
+| Author: fjrodafo - Thu, 23 Jan 2025 15:13:56 +0000
+| 
 * Commit: d032595 - v150
 | Author: fjrodafo - Wed, 22 Jan 2025 23:11:15 +0000
 |